bounding_box/requirements.txt
flask
requests
torch
torchvision
numpy
opencv-python
pillow
mediapipe
facenet-pytorch